Cast VeteranCap to float in WarheadType and WeaponType
Replaced direct assignment of RulesClass::Instance->VeteranCap with a static_cast to float in both WarheadType and WeaponType Body.h files to ensure type consistency.

+++ src/Ext/WarheadType/Body.h
@@ -405,7 +405,7 @@ public:
 
 			, AffectsBelowPercent { 1.0 }
 			, AffectsAbovePercent { 0.0 }
-			, AffectsBelowVeterancy { RulesClass::Instance->VeteranCap }
+			, AffectsBelowVeterancy { static_cast<float>(RulesClass::Instance->VeteranCap) }
 			, AffectsAboveVeterancy { 0.0f }
 			, AffectsNeutral { true }
 			, AffectsGround { true }
