+++ src/Ext/WeaponType/Body.h
@@ -116,7 +116,7 @@ public:
 			, CanTargetHouses { AffectedHouse::All }
 			, CanTarget_MaxHealth { 1.0 }
 			, CanTarget_MinHealth { 0.0 }
-			, CanTarget_MaxVeterancy { RulesClass::Instance->VeteranCap }
+			, CanTarget_MaxVeterancy { static_cast<float>(RulesClass::Instance->VeteranCap) }
 			, CanTarget_MinVeterancy { 0.0f }
 			, Burst_Delays {}
 			, Burst_FireWithinSequence { false }
